+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added `ExecutionWrapper` functional interface which adds AOP like decorating functionality but without annotations.
 - Added `Wait.until` convenience methods for waiting for a specific condition to become true.
 - Added `Threads.waitUntil` convenience methods for waiting for a specific condition to become true which all delegate to `Wait.until`.
+- Added `LoggingFunction` functional interface to be able to pass logging functions for logging purposes.
+- Added `LoggingAdapter` as a minimal logging abstraction that can be implemented for any existing logging framework.
+- Added `Retry.noRetry` method which returns `Retry.NO_RETRY`.
 
 ---